+++ qtfred/include/mainwindow.h
@@ -16,6 +16,9 @@ public:
     explicit MainWindow(QWidget *parent = 0);
     ~MainWindow();
 
+public slots:
+    void loadMission();
+
 private:
     Ui::MainWindow *ui;
 };