+++ qtfred/CHANGELOG.md
@@ -6,3 +6,6 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased][Unreleased]
 ### Added
 - Display editor grid.
+- Load a mission.
+- Handle window resizing.
+- Keyboard controls.